Refactor export controls into component.
