+++ reimplementation/app/src/App.js
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import CardSelector from "./components/CardSelector";
 import Card from "./components/Card";
+import Exporter from "./components/Exporter";
 import Templates from "./data/Cards";
 import "./App.css";
 
@@ -14,6 +15,9 @@ function App() {
       </div>
       <div>
         <Card template={card} />
+      </div>
+      <div>
+        <Exporter />
       </div>
     </div>
   );
