+++ reimplementation/app/src/components/Exporter/style.css
@@ -1,0 +1,3 @@
+div.controls {
+  border: 1px solid black;
+}
